+++ packages/shared/akamai-edgeworker-sdk/CHANGELOG.md
@@ -8,6 +8,12 @@ All notable changes to the LaunchDarkly SDK for Akamai Workers will be documente
   * dependencies
     * @launchdarkly/js-server-sdk-common bumped from ^1.0.2 to ^1.0.3
 
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/js-server-sdk-common bumped from ^1.0.3 to ^1.0.4
+
 ## [0.2.0](https://github.com/launchdarkly/js-core/compare/akamai-edgeworker-sdk-common-v0.1.1...akamai-edgeworker-sdk-common-v0.2.0) (2023-06-09)